+++ fluent_pages/pagetypes/flatpage/migrations/0001_initial.py
@@ -6,6 +6,9 @@ from django.db import models
 
 
 class Migration(SchemaMigration):
+    depends_on = (
+        ("fluent_pages", "0004_remove_sort_order"),
+    )
 
     def forwards(self, orm):
         # Adding model 'FlatPage'