+++ fluent_pages/pagetypes/textfile/migrations/0001_initial.py
@@ -5,6 +5,9 @@ from south.v2 import SchemaMigration
 from django.db import models
 
 class Migration(SchemaMigration):
+    depends_on = (
+        ("fluent_pages", "0004_remove_sort_order"),
+    )
 
     def forwards(self, orm):